requirements.txt
aliyun-python-sdk-core==2.13.4
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-dysmsapi==1.0.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-debug-toolbar==1.11
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.9.4
idna==2.8
jieba==0.39
jmespath==0.9.4
mysqlclient==1.4.1
Pillow==6.0.0
PyMySQL==0.9.3
python-memcached==1.59
pytz==2018.9
qiniu==7.2.4
redis==3.2.1
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
webencodings==0.5.1
Whoosh==2.7.4
